Redesigns Deltastation's white ship into a luxury NT frigate. (#39278)
Delta's white ship has always come off as too large, empty, and boxy to me, personally, so I set out to redesign it into something with an actual shape and more details!

Also, it's infested with starving giant spiders. They have half as much health and damage as their regular versions so they're not a complete nightmare to take out, but they're still dangerous to an unarmed explorer.